% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binseq.R
\name{read_binseq}
\alias{read_binseq}
\alias{readBinseq}
\title{Read a BinSeq file}
\usage{
read_binseq(file)

readBinseq(file)
}
\arguments{
\item{file}{Path to the BinSeq file}
}
\value{
Invisibly returns NULL (the function is used for its side effects of printing information)
}
\description{
This function reads and displays information about a BinSeq file, including
file statistics and a preview of the first few records.
}
\examples{
\dontrun{
# If you have a sample.binseq file:
read_binseq("path/to/sample.binseq")
}
}
