% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binseq.R
\name{test_readBinseq}
\alias{test_readBinseq}
\title{Test reading the sample BinSeq file}
\usage{
test_readBinseq()
}
\value{
Invisibly returns NULL
}
\description{
This function demonstrates how to use the readBinseq function with the
included sample file in the package.
}
\examples{
test_readBinseq()
}
